@@ -1,6 +1,6 @@
 {
   "name": "not-store",
-  "version": "3.1.6",
+  "version": "3.1.7",
   "description": "file storage interface",
   "main": "index.cjs",
   "repository": "https://github.com/interrupter/not-store",
@@ -40,71 +40,71 @@
   "author": "interrupter <alex.repin@mail.ru> (http://github.com/interrupter)",
   "license": "ISC",
   "dependencies": {
-    "@shopify/draggable": "^1.1.3",
-    "aws-sdk": "^2.1692.0",
+    "@shopify/draggable": "^1.2.1",
+    "aws-sdk": "^2.1693.0",
     "download": "^8.0.0",
     "easy-yandex-s3": "^2.0.0",
     "is-stream": "2.*",
-    "jsdoc": "^4.0.4",
+    "jsdoc": "^4.0.5",
     "mkdirp": "*",
-    "not-error": "^0.3.0",
+    "not-error": "^0.3.2",
     "not-path": "^1.0.6",
     "not-validation": "^0.0.11",
-    "sharp": "^0.33.5",
+    "sharp": "^0.34.5",
     "streamifier": "^0.1.1",
     "valid-url": "^1.0.9"
   },
   "devDependencies": {
     "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/plugin-syntax-class-properties": "^7.12.13",
-    "@cypress/code-coverage": "^3.13.7",
+    "@cypress/code-coverage": "^3.14.7",
     "babel-eslint": "^10.1.0",
-    "babel-plugin-istanbul": "^7.0.0",
+    "babel-plugin-istanbul": "^7.0.1",
     "chai": "*",
-    "chai-as-promised": "^8.0.0",
+    "chai-as-promised": "^8.0.2",
     "connect": "^3.7.0",
-    "connect-mongo": "^5.1.0",
-    "cors": "^2.8.5",
-    "cypress": "^13.16.0",
+    "connect-mongo": "^6.0.0",
+    "cors": "^2.8.6",
+    "cypress": "^15.9.0",
     "cypress-file-upload": "^5.0.8",
     "eslint": "*",
-    "eslint-plugin-perfectionist": "^4.0.3",
-    "express": "^4.21.1",
-    "express-fileupload": "^1.5.1",
-    "express-session": "^1.18.1",
-    "helmet": "^8.0.0",
+    "eslint-plugin-perfectionist": "^5.4.0",
+    "express": "^5.2.1",
+    "express-fileupload": "^1.5.2",
+    "express-session": "^1.19.0",
+    "helmet": "^8.1.0",
     "husky": "^9.1.7",
     "ink-docstrap": "^1.3.2",
-    "ioredis": "^5.4.1",
+    "ioredis": "^5.9.2",
     "method-override": "^3.0.0",
     "mocha": "*",
     "mockgoose": "^8.0.4",
-    "mongodb-memory-server": "^10.1.2",
-    "mongoose": "^8.8.2",
+    "mongodb-memory-server": "^11.0.1",
+    "mongoose": "^9.1.5",
     "not-bulma": "^2.0.2",
     "not-config": "^0.1.8",
     "not-log": "^0.0.21",
     "not-meta": "^0.0.28",
-    "not-node": "^6.5.37",
+    "not-node": "^6.5.45",
     "not-options": "^0.3.0",
     "not-user": "^1.0.2",
     "nyc": "^17.1.0",
-    "postcss": "^8.4.49",
-    "redis": "^4.7.0",
+    "postcss": "^8.5.6",
+    "redis": "^5.10.0",
     "request": "^2.88.2",
-    "retire": "^5.2.5",
+    "retire": "^5.4.2",
     "@rollup/plugin-babel": "*",
     "@rollup/plugin-commonjs": "*",
     "rollup-plugin-filesize": "^10.0.0",
     "rollup-plugin-istanbul": "^5.0.0",
     "@rollup/plugin-node-resolve": "*",
     "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-svelte": "^7.2.2",
-    "superagent": "^10.1.1",
-    "svelte": ">=5.2.7",
-    "svelte-dnd-action": "^0.9.52",
+    "rollup-plugin-svelte": "^7.2.3",
+    "superagent": "^10.3.0",
+    "svelte": ">=5.49.1",
+    "svelte-dnd-action": "^0.9.69",
     "taffydb": "^2.7.3",
-    "uuid": "^11.0.3"
+    "uuid": "^13.0.0"
   },
   "nyc": {
     "include": [
